Move KVValue and Range to key_value module
